safari-600.7-branch/LayoutTests/js/dfg-arguments-alias-escape-expected.txt

Tests aliased uses of 'arguments' that escape.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(42) is 1764 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
